chore: move api to appropriate folders
