+++ src/main/java/model/api/boxoffice/BoxOfficeSystemInterface.java
@@ -1,9 +1,6 @@
-package api;
+package model.api.boxoffice;
 
-import entities.Booking;
 
-import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
